fix: use bigint for competition ranking fields and include stream error details in upload handler

+++ competition_ranking_scores.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS competition_ranking_scores (
     id bigserial PRIMARY KEY,
     pid numeric(10) NOT NULL,
-    festival_id integer NOT NULL,
-    score integer NOT NULL,
+    festival_id bigint NOT NULL,
+    score bigint NOT NULL,
     team_id smallint NOT NULL,
-    team_score integer NOT NULL,
+    team_score bigint NOT NULL,
     is_first_upload boolean NOT NULL
 );
